+++ epam-promo/components/buttons/Button.tsx
@@ -2,13 +2,13 @@ import { Button as uuiButton, ButtonProps as UuiButtonProps, ButtonMode } from '
 import { withMods } from '@epam/uui-core';
 import { FillStyle } from '../types';
 
-export type ButtonColor = 'blue' | 'green' | 'red' | 'gray50';
-export const allButtonColors: ButtonColor[] = [
-    'blue', 'green', 'red', 'gray50',
-];
+export type ButtonColor = 'blue' | 'green' | 'red' | 'gray50' | 'gray';
 
 export interface ButtonMods {
     fill?: FillStyle;
+    /** Button color.
+     * Note that 'gray50' is deprecated and will be removed in future versions, please use 'gray' instead.
+     * */
     color?: ButtonColor;
 }