+++ loveship/components/overlays/Tooltip.tsx
@@ -3,10 +3,10 @@ import { Tooltip as uuiTooltip, TooltipProps as UuiTooltipProps } from '@epam/uu
 
 export interface TooltipMods {
     /**
-     * Tooltip color
-     * Note that 'night900' is deprecated and will be removed in future versions, please use 'night800' instead.
+     * Tooltip color.
+     * Note that 'night900' is deprecated and will be removed in future versions, please use 'gray' instead.
      */
-    color?: 'white' | 'fire' | 'night800' | 'night900';
+    color?: 'white' | 'fire' | 'gray' | 'night900';
 }
 
 export type TooltipProps = Omit<UuiTooltipProps, 'color'> & TooltipMods;
@@ -15,6 +15,6 @@ export const Tooltip = withMods< Omit<UuiTooltipProps, 'color'>, TooltipMods>(
     uuiTooltip,
     () => [],
     (props) => ({
-        color: props.color === 'night900' ? 'night800' : props.color ?? 'night800',
+        color: props.color === 'night900' ? 'gray' : props.color ?? 'gray',
     } as TooltipProps),
 );